code refactor, grid view
